Revert "Feature/timer"
