+++ src/components/Timer/timer.css
@@ -1,7 +1,6 @@
 .timer {
-    flex: 1 1;
-    transition: background-color 200ms ease;
-    width: 100%;
+    transition: background-color 500ms ease;
+    width: 400px;
     display: flex;
     flex-direction: column;
     justify-content: center;
@@ -10,84 +9,39 @@
 }
 
 .timer__ticker {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    width: 400px;
-    height: 400px;
-    background-color: #fff;
-    border-radius: 50px;
-    box-shadow: inset 0 0 15px rgba(0, 0, 0, 0.25);
+    cursor: pointer;
     font-family: Roboto, sans-serif;
     font-weight: bold;
     font-size: 72px;
     line-height: 84px;
     color: #fff;
-}
-
-.timer__time {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    border-radius: 50%;
-    width: 300px;
-    height: 300px;
-    left: 362px;
-    top: 210px;
-    border: 12px solid #F55A5A;
-    filter: drop-shadow(3px 3px 5px rgba(0, 0, 0, 0.25));
-    font-weight: bold;
-    font-size: 72px;
-    line-height: 84px;
-    cursor: pointer;
-    transition: all 200ms ease;
-}
-
-.timer__time:hover {
-    opacity: .7;
+    text-shadow: 3px 3px 5px rgba(0, 0, 0, 0.25);
 }
 
 .timer__buttons {
     display: flex;
     align-items: center;
     justify-content: space-between;
-    width: 400px;
-    height: 66px;
-    background: #ffffff;
-    border-radius: 20px;
+    width: 300px;
+    height: 50px;
+    border-radius: 15px;
     overflow: hidden;
-    margin-top: 27px;
-    box-shadow: inset 0 0 15px rgb(0 0 0 / 25%);
 }
 
 .timer__button {
-    flex: 1 0;
     font-family: Roboto, sans-serif;
-    font-size: 22px;
-    color: #F55A5A;
-    font-weight: bold;
-    line-height: 26px;
     cursor: pointer;
+    flex-grow: 1;
     border: none;
     background-color: #fff;
     height: 100%;
-    transition: all 200ms ease;
+    transition: background-color 200ms ease-in-out;
 }
 
 .timer__button:hover {
-    opacity: .7;
-}
-
-.timer__button:focus {
-    outline: none;
+    background-color: rgba(255, 255, 255, .5);
 }
 
 .timer__button_active {
-    color: #FFFFFF;
-    cursor: default;
-    box-shadow: inset 0 0 10px 5px rgba(51, 51, 51, 0.25);
-}
-
-.timer__button_active:hover {
-    opacity: 1;
+    background-color: rgba(255, 255, 255, .3);
 }
