--- src/components/app.css
@@ -1,6 +1,0 @@
-.App {
-    display: flex;
-    justify-content: space-between;
-    align-items: stretch;
-    min-height: 100vh;
-}
